Factor out test-mnt-ns-child test binary
Factor out a new test binary, test-mnt-ns-child. This allows us to
simplify test-mnt-ns to only take care of creating a new mount
namespace, making it more readily reusable.
