+++ dev/build.rs
@@ -688,6 +688,9 @@ fn prepare_test_files() {
     let src = data_dir.join("test-mnt-ns.c");
     cc(&src, "test-mnt-ns.bin", &[]);
 
+    let src = data_dir.join("test-mnt-ns-child.c");
+    cc(&src, "test-mnt-ns-child.bin", &[]);
+
     let src = data_dir.join("test-block.c");
     let ld_script = data_dir.join("test-block-augmented.ld");
     let args = &[
